+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/controllers/RatingController.java
@@ -6,10 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -34,12 +31,12 @@ public class RatingController {
     }
 
     @PostMapping("/rating/validate")
-    public String validate(@Valid Rating rating, BindingResult result, Model model) {
+    public String validate(@Valid @RequestBody Rating rating, BindingResult result, Model model) {
         // TODO: check data valid and save to db, after saving return Rating list
         if (result.hasErrors()) {
             return "rating/add";
         }
-        model.addAttribute("bidList", rating);
+        ratingService.updateRating(rating);
         return "redirect:/rating/list";
     }
 
@@ -51,11 +48,12 @@ public class RatingController {
     }
 
     @PostMapping("/rating/update/{id}")
-    public String updateRating(@PathVariable("id") Integer id, @Valid Rating rating,
+    public String updateRating(@PathVariable("id") Integer id, @Valid @RequestBody Rating rating,
                              BindingResult result) {
-        if (result.hasErrors()) {
-            return "/rating/update/"+id;
+        if (result.hasErrors()|| ratingService.findBidListbyID(id)==null) {
+            return "redirect:/rating/list";
         }
+        rating.setId(id);
         ratingService.updateRating(rating);
         // TODO: check required fields, if valid call service to update Rating and return Rating list
         return "redirect:/rating/list";
@@ -63,6 +61,9 @@ public class RatingController {
 
     @GetMapping("/rating/delete/{id}")
     public String deleteRating(@PathVariable("id") Integer id) {
+        if (ratingService.findBidListbyID(id)==null) {
+            return "redirect:/rating/list";
+        }
         ratingService.deleteRating(id);
         // TODO: Find Rating by Id and delete the Rating, return to Rating list
         return "redirect:/rating/list";
